Pin apache-airflow-providers-slack to 7.3.2
After 7.3.2 it is required to use slack_con_id instead of token for `SlackAPIPostOperator`. This will not work with the current airflow setup as we share the same token for all instances. Also, it will require migrations in the users DAGs. Therefore I temporarily pin the version to 7.3.2 and create a separate task for this.